boards: add Challenger RP2040 LoRa
