+++ src/machine/board_nano-rp2040.go
@@ -13,6 +13,11 @@
 //
 package machine
 
+import (
+	"device/rp"
+	"runtime/interrupt"
+)
+
 // Digital Pins
 const (
 	D2  Pin = GPIO25
@@ -104,3 +109,26 @@ var (
 	usb_VID uint16 = 0x2341
 	usb_PID uint16 = 0x005e
 )
+
+// UART pins
+const (
+	UART0_TX_PIN = GPIO0
+	UART0_RX_PIN = GPIO1
+	UART_TX_PIN  = UART0_TX_PIN
+	UART_RX_PIN  = UART0_RX_PIN
+)
+
+// UART on the RP2040
+var (
+	UART0  = &_UART0
+	_UART0 = UART{
+		Buffer: NewRingBuffer(),
+		Bus:    rp.UART0,
+	}
+)
+
+var DefaultUART = UART0
+
+func init() {
+	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
+}
