+++ src/machine/board_xiao-rp2040.go
@@ -9,6 +9,11 @@
 //
 package machine
 
+import (
+	"device/rp"
+	"runtime/interrupt"
+)
+
 // Digital Pins
 const (
 	D0  Pin = GPIO26
@@ -66,6 +71,29 @@ const (
 const (
 	xoscFreq = 12 // MHz
 )
+
+// UART pins
+const (
+	UART0_TX_PIN = GPIO0
+	UART0_RX_PIN = GPIO1
+	UART_TX_PIN  = UART0_TX_PIN
+	UART_RX_PIN  = UART0_RX_PIN
+)
+
+// UART on the RP2040
+var (
+	UART0  = &_UART0
+	_UART0 = UART{
+		Buffer: NewRingBuffer(),
+		Bus:    rp.UART0,
+	}
+)
+
+var DefaultUART = UART0
+
+func init() {
+	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
+}
 
 // USB CDC identifiers
 const (
