+++ src/machine/machine_rp2040.go
@@ -5,7 +5,6 @@ package machine
 
 import (
 	"device/rp"
-	"runtime/interrupt"
 )
 
 const deviceName = rp.Device
@@ -89,38 +88,6 @@ func ticks() uint64 {
 //go:linkname lightSleep runtime.machineLightSleep
 func lightSleep(ticks uint64) {
 	timer.lightSleep(ticks)
-}
-
-// UART pins
-const (
-	UART_TX_PIN  = UART0_TX_PIN
-	UART_RX_PIN  = UART0_RX_PIN
-	UART0_TX_PIN = GPIO0
-	UART0_RX_PIN = GPIO1
-	UART1_TX_PIN = GPIO8
-	UART1_RX_PIN = GPIO9
-)
-
-// UART on the RP2040
-var (
-	UART0  = &_UART0
-	_UART0 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART0,
-	}
-
-	UART1  = &_UART1
-	_UART1 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    rp.UART1,
-	}
-)
-
-var DefaultUART = UART0
-
-func init() {
-	UART0.Interrupt = interrupt.New(rp.IRQ_UART0_IRQ, _UART0.handleInterrupt)
-	UART1.Interrupt = interrupt.New(rp.IRQ_UART1_IRQ, _UART1.handleInterrupt)
 }
 
 // CurrentCore returns the core number the call was made from.
