+++ targets/challenger-rp2040.json
@@ -1,0 +1,11 @@
+{
+    "inherits": [
+        "rp2040"
+    ],
+    "serial": "uart",
+    "build-tags": ["challenger_rp2040"],
+    "linkerscript": "targets/feather-rp2040.ld",
+    "extra-files": [
+        "targets/feather-rp2040-boot-stage2.S"
+    ]
+}
